requirements.txt
streamlit
supabase
PyPDF2
python-dotenv
pandas==2.1.0
